изменён тип переменных

+++ WindowsFormsApplication2/Зубарева ПИ2-20.cs
@@ -33,9 +33,9 @@ namespace WindowsFormsApplication2
 
         private void button4_Click(object sender, EventArgs e)
         {
-            int a = Convert.ToInt32(textBox1.Text);
-            int b = Convert.ToInt32(textBox2.Text);
-            int c;
+            float a = Convert.ToInt32(textBox1.Text);
+            float b = Convert.ToInt32(textBox2.Text);
+            float c;
             c = a / b;
             textBox3.Text = Convert.ToString(c);
             label4.Text = "Деление";
@@ -49,9 +49,9 @@ namespace WindowsFormsApplication2
 
         private void button1_Click(object sender, EventArgs e)
         {
-            int a = Convert.ToInt32(textBox1.Text);
-            int b = Convert.ToInt32(textBox2.Text);
-            int c;
+            float a = Convert.ToInt32(textBox1.Text);
+            float b = Convert.ToInt32(textBox2.Text);
+            float c;
             c = a + b;
             textBox3.Text = Convert.ToString(c);
             label4.Text = "Сложение";
@@ -60,9 +60,9 @@ namespace WindowsFormsApplication2
 
         private void button2_Click(object sender, EventArgs e)
         {
-            int a = Convert.ToInt32(textBox1.Text);
-            int b = Convert.ToInt32(textBox2.Text);
-            int c;
+            float a = Convert.ToInt32(textBox1.Text);
+            float b = Convert.ToInt32(textBox2.Text);
+            float c;
             c = a - b;
             textBox3.Text = Convert.ToString(c);
             label4.Text = "Вычитание";
@@ -71,9 +71,9 @@ namespace WindowsFormsApplication2
 
         private void button3_Click(object sender, EventArgs e)
         {
-            int a = Convert.ToInt32(textBox1.Text);
-            int b = Convert.ToInt32(textBox2.Text);
-            int c;
+            float a = Convert.ToInt32(textBox1.Text);
+            float b = Convert.ToInt32(textBox2.Text);
+            float c;
             c = a * b;
             textBox3.Text = Convert.ToString(c);
             label4.Text = "Умножение";
